Move `AsyncReadBody` to `fs`
So it can be used by the upcoming `ServeFile`
